+++ fabric-loader-shim/src/main/java/net/fabricmc/loader/api/FabricLoader.java
@@ -34,6 +34,8 @@ public interface FabricLoader {
 
     EnvType getEnvironmentType();
 
+    String getRawGameVersion();
+
     @Deprecated
     Object getGameInstance();